backend/pyproject.toml
[project]
name = "emailai"
version = "0.1.0"
description = ""
authors = [
    {name = "Darcy Anderson",email = "darcy.anderson@teamhgs.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "azure-functions>=1.23.0,<2.0.0",
    "azure-cosmos (>=4.9.0,<5.0.0)",
    "python-dotenv>=1.0.0,<2.0.0",
    "html2text (>=2025.4.15,<2026.0.0)",
    "aihub @ file:./packages/aihub-0.0.1-py3-none-any.whl",
    "pydantic (>=2.11.7,<3.0.0)",
    "msal (>=1.33.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.16"
